backend/pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "A customer support agent that answers FAQ's and raises support ticket"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "typing-extensions>=4.15.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.38",
    "sentence-transformers>=2.2.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "markdown>=3.5.0",
    "pydantic-settings>=2.13.1",
    "loguru>=0.7.3",
]
